example/requirements.txt
Django
South
distribute
django-appconf
django-grappelli